fixed dependencies issues
